% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmodels.R
\name{cuts}
\alias{cuts}
\title{Wrapper to the cut() function}
\usage{
cuts(data, npieces, simple.labels = T, ...)
}
\description{
A wrapper to the cut() function, so that you can automatically break into quantiles as
the default behavior, otherwise if the breakpoints are included, then just break on those.
In all cases, include.lowest is set to True
}
