% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmodels.R
\name{lcmodels}
\alias{lcmodels}
\title{Risk Predictions from Lung Cancer Models}
\usage{
lcmodels(x)
}
\arguments{
\item{x}{A data frame or matrix containing individuals' covariate values.
 Covariates should be in the following columns and numerical formats:

 \itemize{
 \item column 1 - current age (numeric);
 \item column 2 - gender (1=Female, 0=Male);
 \item column 3 - years smoked (numeric);
 \item column 4 - years quit (numeric with 0 to indicate current smoker);
 \item column 5 - cigarettes per day (numeric);
 \item column 6 - race (0=Non-hispanic white,
                  1=Non-hispanic Black/African American, 
                  2=Hispanic, 
                  3=Other Ethnicity);
 \item column 7 - lung disease (1=COPD or Emphysema, 0=No COPD or Emphysema);
 \item column 8 - number of first degree relatives with lung cancer (0,1,2);
 \item column 9 - bmi;
 \item column 10 - highest education level (1=<12 grade, 
                                      2=HS graduate, 
                                      3=post hs, no college, 
                                      4=associate degree/some college, 
                                      5=bachelors degree, 
                                      6=graduate school);
 \item column 11 - asbestos exposure binary indicator;
 \item column 12 - prior history of pneumonia indicator;
 \item column 13 - prior history of cancer indicator;
 \item column 14 - family history of lung cancer (0=none, 1=early onset, 2=late onset);
 \item column 15 - Dust exposure  (binary indicator);
 \item column 16 - 2 or more first degree relatives with cancer (binary indicator);
 \item column 17 - 1 or more first degree relatives with smoking cancer (binary indicator);
 \item column 18 - no hay fever (binary indicator);
 \item column 19 - asian ethnicity (binary indicator);
 \item column 20 - islander ethnicity (binary indicator);
 \item column 21 - American indian ethnicity (binary indicator);
 }
 
 Risk factors used by each model
 \tabular{cccccccccc}{
 Covariate \tab Bach \tab Spitz \tab LLP \tab Hoggart \tab PLCOm2012 \tab Pittsburgh \tab LLPi \tab LCRAT \tab LCDRAT\cr
 Age \tab Yes \tab Yes \tab Yes \tab Yes \tab Yes \tab Yes \tab Yes \tab Yes \tab Yes \cr
 Gender \tab Yes \tab Yes \tab Yes \tab No \tab No \tab No \tab Yes \tab Yes \tab Yes \cr
 Race/ethnicity \tab No \tab No \tab No \tab No \tab Yes \tab No \tab No \tab Yes \tab Yes\cr
 Asian \tab No \tab No \tab No \tab No \tab Yes \tab No \tab No \tab No \tab No\cr
 Pacific islander \tab No \tab No \tab No \tab No \tab Yes \tab No \tab No \tab No \tab No\cr
 America Indian \tab No \tab No \tab No \tab No \tab Yes \tab No \tab No \tab No \tab No\cr
 Education \tab No \tab No \tab No \tab No \tab Yes \tab No \tab No \tab Yes \tab Yes\cr 
 BMI \tab No \tab No \tab No \tab No \tab Yes \tab No \tab No \tab Yes \tab Yes\cr
 Smoking status \tab No \tab Yes \tab No \tab Yes \tab Yes \tab Yes \tab No \tab No \tab No\cr
 Years/Age quit \tab Yes \tab Yes \tab No \tab Yes \tab Yes \tab No \tab No \tab Yes \tab Yes\cr
 Years smoked \tab Yes \tab No \tab Yes \tab Yes \tab Yes \tab Yes \tab Yes \tab Yes \tab Yes\cr
 Cigs per day \tab Yes \tab No \tab No \tab Yes \tab Yes \tab Yes \tab No \tab Yes \tab Yes\cr
 Pack-years \tab No \tab Yes \tab No \tab No \tab No \tab No \tab No \tab Yes \tab Yes\cr
 Prior cancer \tab No \tab No \tab Yes \tab No \tab Yes \tab No \tab Yes \tab No \tab No\cr
 Lung disease \tab No \tab Yes \tab No \tab No \tab Yes \tab No \tab Yes \tab Yes \tab Yes\cr
 Pneumonia \tab No \tab No \tab Yes \tab No \tab No \tab No \tab No \tab No \tab No\cr
 Hayfever \tab No \tab Yes \tab No \tab No \tab No \tab No \tab No \tab No \tab No\cr
 Asbestos exposure \tab Yes \tab Yes \tab Yes \tab No \tab No \tab No \tab No \tab No \tab No\cr
 Dust exposure \tab No \tab Yes \tab No \tab No \tab No \tab No \tab No \tab No \tab No\cr
 Any FDR w/ cancer \tab No \tab Yes \tab No \tab No \tab No \tab No \tab No \tab No \tab No\cr
 Any FDR w/ smoking-related cancer \tab No \tab Yes \tab No \tab No \tab No \tab No \tab No \tab No \tab No\cr
 Any FDR w/ LC \tab No \tab No \tab Yes \tab No \tab Yes \tab No \tab Yes \tab Yes \tab Yes\cr
 Num. FDR w/ LC \tab No \tab No \tab No \tab No \tab No \tab No \tab No \tab Yes \tab Yes\cr
 Early onset FDR w/ LC \tab No \tab No \tab Yes \tab No \tab No \tab No \tab Yes \tab No \tab No
 }}
}
\value{
A numeric matrix containing individuals' predictions:

 \itemize{
 \item column 1 - An indicator variable for whether the individual is eligible 
             for CT lung screening according to 
             US Preventive Services Task Force (USPSTF) recommendations.
 \item column 2 - This is the probability of dying from lung cancer 
             within 5 years if not undergoing screening (Katki, 2016).
 \item column 3 - This is the reduction in the probability of dying from lung cancer in 5 years 
 \item column 4 - This is the probability of being diagnosed with lung cancer
             within 5 years if not undergoing screening (Katki, 2016).
 \item column 5 - This is the extra probability of lung cancer diagnosis in 5 years
             if undergoing 3 yearly CT lung screens as in the NLST  (Katki, 2016).
 \item column 6 - This is the probability of having at least one false-positive 
             CT screen out of 3 screens (Katki, 2016).
 \item column 7 - This is the expected number of false-positive CT screens 
              after 3 screens (Katki, 2016).
 \item column This is the probability of being diagnosed with lung cancer 
             within 10 years if not undergoing screening (Bach, 2003).
 \item column 9 - This is the probability of being diagnosed with lung cancer 
             within 1 years if not undergoing screening (Hoggart, 2012).
 \item column 10 - This is the probability of being diagnosed with lung cancer 
             within 5 years if not undergoing screening (LLP, 2008).
 \item column 11 - This is the probability of being diagnosed with lung cancer 
             within 8.7 years if not undergoing screening (LLPi, 2015).
 \item column 12 - This is the probability of being diagnosed with lung cancer 
             within 1 years if not undergoing screening (Spitz, 2007).
 \item column 13 - This is the probability of being diagnosed with lung cancer 
             within 6 years if not undergoing screening (Tammemagi, 2013).
 \item column 14 - This is the probability of being diagnosed with lung cancer
             within 6 years if not undergoing screening (Pittsburgh, 2015).            
 }
}
\description{
The R package provides individual risks of lung cancer and lung cancer death 
based on various published papers: Bach et al., 2003; Spitz et al., 2007; 
Cassidy et al., 2008 (LLP); Hoggart et al., 2012; Tammemagi et al., 2013 (PLCOm2012); 
Marcus et al., 2015 (LLPi); Wilson and Weissfeld, 2015 (Pittsburgh);
Katki et al., 2016 (LCRAT and LCDRAT).
}
\section{Warning}{

 VGAM is a required dependency of this package.
 VGAM may automatically be installed the first time this package is used.
 Inputs must be in numerical format to ensure correct output.
 For data frame x, this can be checked using sapply(x,class)
}

\examples{
age <- c(66,58,75,72,56)
bmi <- c(23,28,26,27,24)
cpd <- c(36,36,40,24,40)
emp <- c(0,1,1,0,1)
fam.lung.trend <- c(0,2,0,2,0)
female <- c(0,1,0,1,0)
smkyears <- c(43,37,45,42,29)
qtyears <- c(0,0,9,6,6)
race <- c(0,1,2,2,3)
edu6 <- c(3,5,4,5,5)
asb <- c(0,0,0,0,0)
pneu <- c(0,0,0,0,0)
prior.cancer <- c(0,0,0,0,0)
fam.cancer.onset <- c(0,1,0,2,0)
dust <- c(0,0,0,0,0)
fam.cancer <- c(0,1,0,1,0)
fam.smoke.cancer <- c(0,1,0,1,0)
no.hayfever <- c(1,1,1,1,1)
asian <- c(0,0,0,0,1)
islander <- c(0,0,0,0,0)
indian <- c(0,0,0,0,0)

persons <- data.frame(age,
                      female,
                      smkyears,
                      qtyears,
                      cpd,
                      race,
                      emp,
                      fam.lung.trend,
                      bmi,
                      edu6,
                      asb,
                      pneu,
                      prior.cancer,
                      fam.cancer.onset,
                      dust,
                      fam.cancer,
                      fam.smoke.cancer,
                      no.hayfever,
                      asian,
                      islander,
                      indian)

persons_predictions <- lcmodels(persons)
persons_predictions
}
\references{
\itemize{
 \item Bach PB, Kattan MW, Thornquist MD, et al. 
       Variations in lung cancer risk among smokers. J Natl Cancer Inst 2003;95:470-8.
 \item Spitz MR, Hong WK, Amos CI, et al. 
       A risk model for prediction of lung cancer. J Natl Cancer Inst 2007;99:715-26.
 \item Cassidy A, Myles JP, van Tongeren M, et al. 
       The LLP risk model: an individual risk prediction model for lung cancer. 
       Br J Cancer 2008;98:270-6.
 \item Hoggart C, Brennan P, Tjonneland A, et al. 
       A risk model for lung cancer incidence. Cancer Prev Res (Phila) 2012;5:834-46.
 \item Tammemagi MC, Katki HA, Hocking WG, et al. 
       Selection criteria for lung-cancer screening. N Engl J Med 2013;368:728-36.
 \item Marcus MW, Chen Y, Raji OY, Duffy SW, Field JK.
       LLPi: Liverpool lung project risk prediction model for lung cancer incidence.
       Cancer Prev Res (Phila) 2015;8:570-5.
 \item Wilson DO, Weissfeld J.
       A simple model for predicting lung cancer occurrence in a lung cancer screening program:
       the Pittsburgh predictor. Lung Cancer 2015;89:31-37.      
 \item Katki HA, Kovalchik SA, Berg CD, Cheung LC, Chaturvedi AK.
       Development and validation of risk models to select ever-smokers
       for CT lung cancer screening. JAMA. 2016;315:2300-11.
       doi: 10.1001/jama.2016.6255.
 \item Katki HA, Kovalchik SA, Petito LC, Cheung LC, Jacobs E, Jemal A, Berg CD, Chaturvedi AK.
       Implications of nine risk prediction models for selecting ever-smokers for computed
       tomography lung cancer screening.  Ann Intern Med. 2018;doi: 10.7326/M17-2701. 
          
}
}
\author{
Li C. Cheung, \email{li.cheung@nih.gov}, Stephanie A. Kovalchik, Hormuzd A. Katki
}
